feat: Refactor configuration handling and add config command for documentation
